Moved Notification and ReferentieLijst Service mixins for tests
