+++ src/openzaak/components/autorisaties/tests/admin/test_autorisaties_admin.py
@@ -21,9 +21,9 @@ from openzaak.components.catalogi.tests.factories import (
     InformatieObjectTypeFactory,
     ZaakTypeFactory,
 )
+from openzaak.notifications.tests.mixins import NotificationServiceMixin
 from openzaak.tests.utils import mock_nrc_oas_get
 from openzaak.utils import build_absolute_url
-from openzaak.utils.tests import NotificationServiceMixin
 
 from ...constants import RelatedTypeSelectionMethods
 from ..factories import ApplicatieFactory, AutorisatieSpecFactory
