+++ src/openzaak/components/besluiten/tests/test_notifications_send.py
@@ -13,8 +13,9 @@ from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectFactory,
 )
 from openzaak.notifications.models import FailedNotification
+from openzaak.notifications.tests.mixins import NotificationServiceMixin
 from openzaak.notifications.tests.utils import LOGGING_SETTINGS
-from openzaak.utils.tests import JWTAuthMixin, NotificationServiceMixin
+from openzaak.utils.tests import JWTAuthMixin
 
 from ..constants import VervalRedenen
 from .factories import BesluitFactory, BesluitInformatieObjectFactory
