+++ src/openzaak/components/catalogi/tests/admin/test_resultaattype_admin.py
@@ -14,7 +14,8 @@ from openzaak.selectielijst.tests import (
     mock_resource_get,
     mock_resource_list,
 )
-from openzaak.utils.tests import ClearCachesMixin, ReferentieLijstServiceMixin
+from openzaak.selectielijst.tests.mixins import ReferentieLijstServiceMixin
+from openzaak.utils.tests import ClearCachesMixin
 
 from ...models import ResultaatType
 from ..factories import ResultaatTypeFactory, ZaakTypeFactory
