+++ src/openzaak/components/zaken/tests/test_notifications_send.py
@@ -31,7 +31,8 @@ from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectFactory,
 )
 from openzaak.notifications.models import FailedNotification
-from openzaak.utils.tests import JWTAuthMixin, NotificationServiceMixin
+from openzaak.notifications.tests.mixins import NotificationServiceMixin
+from openzaak.utils.tests import JWTAuthMixin
 
 from ..models import Zaak
 from .factories import (
